reorganise into `Product` folder
