--- backend/PriceAlert/src/API/DTOs/Product.cs
@@ -1,9 +1,0 @@
-namespace API.DTOs
-{
-  public class Product
-  {
-    public required int Id { get; set; }
-    public required string Url { get; set; }
-    public required string Name { get; set; }
-  }
-}
