+++ backend/PriceAlert/src/API/Product/Product.cs
@@ -1,0 +1,8 @@
+namespace API.Product;
+
+public class Product
+{
+  public required int Id { get; set; }
+  public required string Url { get; set; }
+  public required string Name { get; set; }
+}
